# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Verspätomat API
  version: 0.4.0
  description: >
    Contract for the Flutter app. Amounts are euro cents; timestamps RFC 3339 UTC; dates YYYY-MM-DD.
    Station ids and trip ids are Transitous ids passed through unchanged (trip ids go in query parameters).
    Identity: one anonymous device per installation. Every route except /health, /v1/devices*,
    /v1/operators, /v1/ngos and /internal/* requires `Authorization: Bearer <token>`.
servers:
  - url: http://127.0.0.1:8080
paths:
  /health:
    get: { summary: Liveness and database check, responses: { "200": { description: "{ok, db}" } } }

  /v1/devices:
    post: { summary: "Create a device; returns {device_id, token}. Store the token in the Keychain.", responses: { "200": { description: device } } }
  /v1/devices/recover:
    post:
      summary: Re-attach a new phone with the recovery code; returns a fresh {device_id, token}.
      requestBody: { content: { application/json: { schema: { type: object, required: [recovery_code], properties: { recovery_code: { type: string } } } } } }
      responses: { "200": { description: device }, "404": { description: unknown code } }

  /v1/stations/nearby:
    get:
      summary: "Rail stations near a point: the phone's fix, or a Stellwerk override for this customer, never a default. Answers {stations, source: gps|stellwerk|none, label}."
      parameters: [{ name: lat, in: query, schema: { type: number } }, { name: lon, in: query, schema: { type: number } }]
      responses: { "200": { description: "[{id, name, lat, lon, distance_m}]" } }
  /v1/stations/search:
    get:
      parameters: [{ name: q, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: "[Station]" } }
  /v1/stations/{id}/departures:
    get:
      summary: Rail departures with live delays, platform, agency, our operator and claims desk, trip id.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: "[{trip_id, line, train_number, headsign, agency_name, operator, desk, category, planned_departure, live_departure, delay_min, platform, cancelled, realtime}]" } }
  /v1/trips:
    get:
      summary: One trip with stop-by-stop scheduled and live times.
      parameters: [{ name: trip_id, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: "{trip_id, line, headsign, agency_name, operator, category, cancelled, stops:[{stop_id, name, scheduled_arrival, live_arrival, scheduled_departure, live_departure, cancelled}]}" }, "404": { description: not found } }
  /v1/operators:
    get: { summary: Operator directory (name, aliases, desk, postal_address, email, accepts_email, last_verified), responses: { "200": { description: "[Operator]" } } }
  /v1/ngos:
    get: { summary: Partner NGOs with confirmed_total_cents and submitted_total_cents, responses: { "200": { description: "[Ngo]" } } }
  /v1/badges:
    get: { summary: Badge catalogue with earned_on for this customer, responses: { "200": { description: "[Badge]" } } }

  /v1/me:
    get: { summary: "Profile: nickname, relay_address, personal_data?, settings, points_total, points_this_week, level_name, next_level_name, next_level_at, home_station", responses: { "200": { description: Customer } } }
    patch:
      summary: Change settings (nickname, ticket, ngo_id, location_mode, notifications, show_on_boards, keep_correspondence, traewelling_linked, onboarding_done, home_station_id/name, muted_stations, nudge_enabled, quiet_from/quiet_to as "HH:MM" or "" to clear).
      responses: { "200": { description: Customer } }
    delete: { summary: Delete the device and everything it owns, responses: { "200": { description: "{deleted}" } } }
  /v1/me/personal-data:
    put:
      summary: Name, address, private e-mail, ticket number, first_class. Assigns the relay address on first call.
      requestBody: { content: { application/json: { schema: { type: object, required: [name, address, email], properties: { name: { type: string }, address: { type: string }, email: { type: string }, ticket_number: { type: string }, first_class: { type: boolean } } } } } }
      responses: { "200": { description: Customer } }
  /v1/me/recovery-code:
    get: { summary: Issue a new recovery code (twelve words). Shown once; the previous code stops working., responses: { "200": { description: "{recovery_code}" } } }
  /v1/me/export:
    get: { summary: Everything about this customer as JSON, responses: { "200": { description: export } } }
  /v1/me/geofence:
    get:
      summary: "The stations the phone should watch in the background (docs/15): frequent check-in stations of the last 30 days (home station always, muted excluded, at most 15) with coordinates stored at check-in; enabled = location_mode == always && nudge_enabled."
      responses: { "200": { description: "{enabled, stations: [{id, name, lat, lon, checkins}], quiet_from, quiet_to}" } }
  /v1/me/push-token:
    put:
      summary: Register the device's push token (devices.push_platform, push_token, push_updated_at). Stored only; APNs/FCM delivery is not wired.
      requestBody: { content: { application/json: { schema: { type: object, required: [platform, token], properties: { platform: { type: string, enum: [ios, android] }, token: { type: string } } } } } }
      responses: { "200": { description: "{stored: true, platform}" }, "400": { description: bad platform or token } }
    delete: { summary: Clear the push token, responses: { "200": { description: "{stored: false}" } } }

  /v1/rides:
    get: { summary: Ride history (newest first), responses: { "200": { description: "[Ride]" } } }
    post:
      summary: Check in. The follower polls the trip from now on.
      requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/CheckIn" } } } }
      responses: { "200": { description: "{ride, stops}" }, "409": { description: already riding } }
  /v1/rides/current:
    get: { summary: "Ride in progress: {ride, stops, eta, claim_from_minute, last_polled_at}; or the last arrival within 2 h as {ride, just_arrived:true}", responses: { "200": { description: live ride }, "404": { description: none } } }
  /v1/rides/current/arrival:
    post:
      summary: Finalise now, with a delay or an actual time (no-data case, or the demo control). Creates the incident if 60+ and awards badges.
      requestBody: { content: { application/json: { schema: { type: object, properties: { delay_minutes: { type: integer }, actual_arrival: { type: string, format: date-time }, cancelled: { type: boolean }, self_entered: { type: boolean } } } } } }
      responses: { "200": { description: "{ride, incident?, bundle_ready, new_badge?}" } }
  /v1/rides/current/dismiss:
    post: { summary: Acknowledge the arrival summary, responses: { "200": { description: ok } } }
  /v1/rides/nachtrag:
    post:
      summary: Record a forgotten ride from a trip id (1 point, claimable if the feed shows 60+, never ranks).
      requestBody: { content: { application/json: { schema: { type: object, required: [trip_id, from_station_id, from_station_name, exit_station_id, exit_station_name], properties: { trip_id: { type: string }, from_station_id: { type: string }, from_station_name: { type: string }, exit_station_id: { type: string }, exit_station_name: { type: string } } } } } }
      responses: { "200": { description: "{ride, incident?}" } }

  /v1/incidents:
    get: { summary: "Ledger: {incidents, claims, summary:{desks:[{desk, open_cents, ready, missing_cents, incident_ids}], ready_desk, confirmed_cents, submitted_cents, capped_cents, oldest_open:{id,line,date,deadline,days_left,warn_from}, min_payout_cents, dticket_monthly_cap_cents}}. Reading refreshes statuses: expiry at the legal deadline, bundle readiness, and the 25 % monthly Deutschlandticket cap (incidents beyond it are `gedeckelt`; capped_cents sums them).", responses: { "200": { description: ledger } } }
  /v1/claims:
    get: { summary: "Claim: {id, desk, ngo_id, status: draft|sent|question|accepted|rejected|bounced, sent_at, expected_reply_by, amount_claimed_cents, amount_confirmed_cents, closed_at, pdf_url}. The scanner's reply nudge (claims.nudged_at) reaches the app as the SSE event claim {nudge: true}, not as a field.", responses: { "200": { description: "[Claim]" } } }
  /v1/claims/draft:
    post:
      summary: Start a claim for one desk (all open incidents there unless incident_ids given). Refused below the 4 € minimum.
      requestBody: { content: { application/json: { schema: { type: object, required: [desk], properties: { desk: { type: string }, incident_ids: { type: array, items: { type: string, format: uuid } } } } } } }
      responses: { "200": { description: "claim + incidents + desk_address, desk_email, desk_accepts_email, personal_data_required, relay_address, needs_recovery_code" }, "412": { description: bundle not ready } }
  /v1/claims/{id}:
    patch:
      summary: "Change the NGO for this claim; set attachments [{upload_id, label}] (one ticket image per month covered)."
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: claim } }
  /v1/uploads:
    post:
      summary: "multipart/form-data: kind (ticket|signature|postal_reply), file. Max 8 MB."
      responses: { "200": { description: "{upload_id, kind, content_type, size}" } }
  /v1/claims/{id}/pdf:
    get:
      summary: The filled EU claim form (Typst) as it stands, draft or sent; the same file is attached to the claim mail as EU-Antrag.pdf. Claim JSON carries it as pdf_url.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: PDF, content: { application/pdf: { schema: { type: string, format: binary } } } }, "404": { description: not found } }
  /v1/claims/{id}/sign:
    post:
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { content: { application/json: { schema: { type: object, required: [typed_name], properties: { typed_name: { type: string }, signature_upload_id: { type: string, format: uuid } } } } } }
      responses: { "200": { description: claim } }
  /v1/claims/{id}/send:
    post:
      summary: The relay. Refused unless signed and personal data present; max 5 per day. Real SMTP when SMTP_URL is set, else dry-run. BCC to the customer. Incidents become eingereicht.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: "{claim, mail}" }, "412": { description: not signed / no personal data / desk takes no e-mail }, "429": { description: daily limit } }

  /v1/mails:
    get: { summary: All mails in and out (newest first), responses: { "200": { description: "[Mail]" } } }
  /v1/mails/{id}/reply:
    post:
      summary: Answer a railway's question through the customer's relay address.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { content: { application/json: { schema: { type: object, required: [body], properties: { body: { type: string } } } } } }
      responses: { "200": { description: Mail } }
  /internal/inbound-mail:
    post:
      summary: Mail-provider webhook (and the demo's "Antwort simulieren"). Accepts our own shape or Postmark's inbound payload as posted (To/ToFull, From/FromFull, Subject, TextBody, MessageID, Headers, base64 Attachments; a RawEmail field is parsed as MIME instead). Matches by relay address, classifies, forwards to the customer, updates statuses. Guarded by ?secret= when INBOUND_SECRET is set.
      requestBody: { content: { application/json: { schema: { type: object, required: [to, from, subject, body], properties: { to: { type: string }, from: { type: string }, subject: { type: string }, body: { type: string }, message_id: { type: string }, in_reply_to: { type: string }, claim_id: { type: string, format: uuid } } } } } }
      responses: { "200": { description: "{mail, outcome, claim_id}" } }
  /internal/inbound-mail/raw:
    post:
      summary: The same webhook taking the raw RFC 822 message as the body (mail-parser). From, To (the address on our domain), Subject, text body (HTML stripped as fallback), Message-ID, In-Reply-To and attachments are extracted; attachments are stored as uploads of kind `inbound` and listed in the mail's attachments. Guarded by ?secret= when INBOUND_SECRET is set.
      requestBody: { content: { message/rfc822: { schema: { type: string } } } }
      responses: { "200": { description: "{mail, outcome, claim_id}" }, "400": { description: not parseable } }

  /v1/community:
    get: { summary: "{minutes, submitted_cents, confirmed_cents, users, ngos:[{id, name, confirmed_cents, submitted_cents}]}", responses: { "200": { description: community } } }
  /v1/boards:
    get:
      summary: "Seven-day board: points of location-verified rides finalised in the last seven days, per customer with show_on_boards. Scope line = rides on my most frequent line of the last 30 days; city = rides starting at a station whose first word matches my home station's; germany = all; a scope with nothing to narrow on falls back to all. Filled from board_seed (seed: true) up to ten entries; the requesting customer always appears with is_me."
      parameters: [{ name: scope, in: query, schema: { type: string, enum: [line, city, germany] } }]
      responses: { "200": { description: "[{rank, name, points, is_me, seed}]" } }

  # Server-sent events on GET /v1/events also carry the deadline scanner's output (hourly):
  #   incident {incident_id, warning: true, legal_deadline, days_left}   21 days before the deadline, once
  #   incident {expired: [ids]}                                          open incidents past the deadline became verfallen
  #   claim    {claim_id, nudge: true, expected_reply_by}                a sent claim passed its expected reply date without an answer, once
  #   claim    {claim_id, status: accepted, source: ngo_report}          an NGO statement confirmed the transfer

  # Stellwerk admin API (x-admin-token; dev and staging only)
  /admin/customers/{key}:
    delete:
      summary: Forget a customer (id, id prefix, nickname or relay address). Deletes the device row; everything cascades. Refused with 409 when the customer has a relay address and a sent claim (a railway may still write to it) unless ?force=true.
      parameters: [{ name: key, in: path, required: true, schema: { type: string } }, { name: force, in: query, schema: { type: boolean } }]
      responses: { "200": { description: "{forgotten, nickname, sent_claims, trip_overrides_cleared}" }, "404": { description: no such customer }, "409": { description: "relay address and sent claims; repeat with ?force=true" } }
  /admin/customers/{key}/push:
    post:
      summary: "Send a test push to the customer's device; body {text?}. Without APNS_*/FCM_* credentials it is one `push (dry-run)` log line. Answers {customer, nickname, platform, token: bool, configured: bool, result: sent | dry-run | no-token | muted | unregistered, title, body}."
      parameters: [{ name: key, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { text: { type: string } } } } } }
      responses: { "200": { description: result }, "404": { description: no such customer } }
  /admin/scan:
    post: { summary: "Run one deadline-scanner pass now: warnings, expiry across all customers, reply nudges, retention sweep.", responses: { "200": { description: "{today, warned, expired, nudged, retained}" } } }
  /admin/ngos/{id}/report:
    post:
      summary: "Import an NGO's monthly statement: JSON {transfers:[{date, amount_cents | amount \"4,50\", reference, counterparty}]} (CSV text `date,amount,reference,counterparty` is accepted too). Each transfer is matched to a sent/question claim of that NGO with exactly that amount, sent within the 60 days before the transfer date, oldest unmatched first (a claim named in reference/counterparty by claimant or id prefix wins). Matches become accepted (amount_confirmed_cents, closed_at), incidents bestätigt with an audit entry \"ngo report\", the bestaetigt badge is awarded, incident and claim events go to the customer, retention runs. One ngo_reports row per import."
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { transfers: { type: array, items: { type: object, required: [date], properties: { date: { type: string, format: date }, amount_cents: { type: integer }, amount: { type: string }, reference: { type: string }, counterparty: { type: string } } } } } } }, text/csv: { schema: { type: string } } } }
      responses: { "200": { description: "{report_id, ngo_id, rows, matched:[claim ids], unmatched:[{date, amount_cents, reference, counterparty}]}" }, "400": { description: no transfers }, "404": { description: no such ngo } }

components:
  schemas:
    CheckIn:
      type: object
      required: [trip_id, from_station_id, from_station_name, exit_station_id, exit_station_name]
      properties:
        trip_id: { type: string }
        from_station_id: { type: string }
        from_station_name: { type: string }
        exit_station_id: { type: string }
        exit_station_name: { type: string }
        ticket: { type: string, enum: [deutschlandticket, zeitkarte, einzelfahrkarte] }
        location: { type: object, properties: { lat: { type: number }, lon: { type: number }, accuracy_m: { type: number } } }
        from_lat: { type: number, description: "the from-station's own coordinates (not the phone's); feed the geofence set" }
        from_lon: { type: number }